move to renderer/
- update image file path to __static
- update jsx compoents to js
