+++ src/renderer/components/Footer.js
