+++ src/renderer/components/MenuIcon.js
