+++ src/renderer/components/Paginator.js
