+++ src/renderer/components/Player.js
