+++ src/renderer/components/ProjectDetail.js
