+++ src/renderer/components/ProjectManager.js
