+++ src/renderer/components/SearchBar.js
